hacked variable expansion in config file to facilitate split indices
git-svn-id: http://biofs04/svn/bioinf/anfo/trunk@1505 93d9a472-daf7-0310-8f21-d7befa554562

+++ sequence.cc
@@ -29,7 +29,7 @@ static inline bool all_acsii_qscores( const std::string& s )
 
 static inline int sol_to_phred( int sol )
 {
-	return 0.5 + 10.0/log(10.0) * log( 1.0 + exp( sol*log(10.0)/10.0 ) ) ;
+	return (int)( 0.5 + 10.0/log(10.0) * log( 1.0 + exp( sol*log(10.0)/10.0 ))) ;
 }
 
 istream& read_fastq( istream& s, QSequence& qs, bool solexa_scores )
